% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/csound.R
\name{csound}
\alias{csound}
\title{Make Some Noise!}
\usage{
csound(orchestra, score, output = "dac")
}
\arguments{
\item{orchestra}{An \link{orchestra} object or string}

\item{score}{An \link{score} object or string}
}
\description{
Create sounds from orchestra and score. The output can be either audio interface
or a file(.wav, .ogg, etc).
}
\seealso{
\url{http://en.flossmanuals.net/csound/ch014_e-rendering-to-file/}
}

