% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/orchestra.R
\name{orchestra}
\alias{create_orchestra}
\alias{orchestra}
\title{Orchestra class}
\usage{
create_orchestra(`0dbfs` = NULL, ctrlinit = NULL, ftgen = NULL,
  kr = NULL, ksmps = NULL, massign = NULL, nchnls = NULL,
  pgmassign = NULL, pset = NULL, seed = NULL, sr = NULL,
  strset = NULL, instruments = NULL)
}
\arguments{
\item{0dbfs}{Sets the value of 0 decibels using full scale amplitude.}

\item{ctrlinit}{Sets the initial values for a set of MIDI controllers.}

\item{ftgen}{Generate a score function table from within the orchestra.}

\item{kr}{Sets the control rate.}

\item{ksmps}{Sets the number of samples in a control period.}

\item{massign}{Assigns a MIDI channel number to a Csound instrument.}

\item{nchnls}{Sets the number of channels of audio output.}

\item{pgmassign}{Assigns an instrument number to a specified (or all) MIDI program(s).}

\item{pset}{Defines and initializes numeric arrays at orchestra load time.}

\item{seed}{Sets the global seed value.}

\item{sr}{Sets the audio sampling rate.}

\item{strset}{Allows a string to be linked with a numeric value.}

\item{instruments}{Sets instruments.}
}
\description{
Orchestra class
}

