% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rsound-package.R
\docType{package}
\name{rsound-package}
\alias{rsound}
\alias{rsound-package}
\title{An R Interface to Csound}
\description{
Csound is a C library for sound and music, which may be usuful for data sonification. This package provides a simple interface to Csound.
}
\seealso{
\url{http://csound.github.io}
}

