% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{create_score}
\alias{score}
\alias{score_f}
\alias{score_i}
\alias{score_t}
\title{Score class}
\usage{
create_score(sections, func = NULL, tempo = NULL)

score_i(...)

score_t(...)

score_f(...)
}
\description{
Score class
}

